Move workflow context implementation to runtime
